requirements.txt
streamlit
requests
beautifulsoup4
feedparser
google-generativeai
